refactor: Line 의 sections 를 일급 컬렉션으로 분리
