+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -1,13 +1,11 @@
 package nextstep.subway.line.domain;
 
 import nextstep.subway.common.BaseEntity;
-import nextstep.subway.line.exception.DownStationExistedException;
-import nextstep.subway.line.exception.EmptyLineException;
-import nextstep.subway.line.exception.InValidUpStationException;
 import nextstep.subway.station.domain.Station;
 
 import javax.persistence.*;
-import java.util.*;
+import java.util.List;
+import java.util.Objects;
 
 @Entity
 public class Line extends BaseEntity {
@@ -18,8 +16,8 @@ public class Line extends BaseEntity {
     private String name;
     private String color;
 
-    @OneToMany(mappedBy = "line", cascade = {CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true, fetch = FetchType.EAGER)
-    private List<Section> sections = new ArrayList<>();
+    @Embedded
+    private Sections sections = new Sections();
 
     public Line() {
     }
@@ -32,7 +30,7 @@ public class Line extends BaseEntity {
     public Line(String name, String color, Station upStation, Station downStation, int distance) {
         this.name = name;
         this.color = color;
-        sections.add(new Section(this, upStation, downStation, distance));
+        addSection(new Section(this, upStation, downStation, distance));
     }
 
     public void update(Line line) {
@@ -41,44 +39,15 @@ public class Line extends BaseEntity {
     }
 
     public void addSection(Section section) {
-        if (sections.isEmpty()) {
-            sections.add(section);
-            return;
-        }
-        validateAddSection(section);
-        sections.add(section);
+        sections.addSection(section);
     }
 
     public void removeSection(Station station) {
-        validateRemoveSections(station);
-        sections.stream()
-                .filter(it -> it.getDownStation().equals(station))
-                .findFirst()
-                .ifPresent(it -> sections.remove(it));
+        sections.removeSection(station);
     }
 
     public List<Station> getStations() {
-        if (sections.isEmpty()) {
-            return Arrays.asList();
-        }
-
-        List<Station> stations = new ArrayList<>();
-        Station downStation = findUpStation();
-        stations.add(downStation);
-
-        while (downStation != null) {
-            Station finalDownStation = downStation;
-            Optional<Section> nextLineStation = sections.stream()
-                    .filter(it -> it.getUpStation() == finalDownStation)
-                    .findFirst();
-            if (!nextLineStation.isPresent()) {
-                break;
-            }
-            downStation = nextLineStation.get().getDownStation();
-            stations.add(downStation);
-        }
-
-        return stations;
+        return sections.getStations();
     }
 
     public int size() {
@@ -98,7 +67,7 @@ public class Line extends BaseEntity {
     }
 
     public List<Section> getSections() {
-        return sections;
+        return sections.toList();
     }
 
     @Override
@@ -112,47 +81,5 @@ public class Line extends BaseEntity {
     @Override
     public int hashCode() {
         return Objects.hash(id, name, color);
-    }
-
-
-    private void validateAddSection(Section section) {
-        validateUpStation(section.getUpStation());
-        if (getStations().stream().anyMatch(
-                section.getDownStation()::equals)) {
-            throw new DownStationExistedException();
-        }
-    }
-
-    private void validateRemoveSections(Station station) {
-        if (sections.size() <= 1) {
-            throw new EmptyLineException();
-        }
-        validateUpStation(station);
-    }
-
-    private void validateUpStation(Station station) {
-        if (!getLastDownStation().equals(station)) {
-            throw new InValidUpStationException();
-        }
-    }
-
-    private Station getLastDownStation() {
-        return sections.get(sections.size() - 1).getDownStation();
-    }
-
-    private Station findUpStation() {
-        Station downStation = sections.get(0).getUpStation();
-        while (downStation != null) {
-            Station finalDownStation = downStation;
-            Optional<Section> nextLineStation = sections.stream()
-                    .filter(it -> it.getDownStation() == finalDownStation)
-                    .findFirst();
-            if (!nextLineStation.isPresent()) {
-                break;
-            }
-            downStation = nextLineStation.get().getUpStation();
-        }
-
-        return downStation;
     }
 }
