+++ src/main/java/nextstep/subway/line/exception/NotLastStationException.java
@@ -1,9 +1,9 @@
 package nextstep.subway.line.exception;
 
-public class InValidUpStationException extends LineDomainException {
+public class NotLastStationException extends LineDomainException {
     private static String message = "상행역은 하행 종점역이어야 합니다.";
 
-    public InValidUpStationException() {
+    public NotLastStationException() {
         super(message);
     }
 }
